Refactor: 비밀번호 암호화 로직을 도메인 클래스에 캡슐화
