+++ used-trading-domain/src/main/java/org/flab/hyunsb/domain/Encryptor/Encryptor.java
@@ -1,4 +1,4 @@
-package org.flab.hyunsb.application.encryptor;
+package org.flab.hyunsb.domain.Encryptor;
 
 public interface Encryptor {
